refactor(folder): unify persisted schema and remove import cycle
Share the runtime folder schema across content, sync and import/export. Keep serialized IDs and backup keys unchanged while removing the import service dependency on FolderManager.
