+++ AGENTS.md
@@ -38,7 +38,7 @@ Use `package.json`, build configs, manifests and CI to verify command names and 
 | ---------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------ |
 | Shared services and storage  | `src/core/services/`; `StorageKeys` in `src/core/types/common.ts`; sync-backed settings also need `SettingsBackupService.ts` defaults/migrations                                     |
 | Feature logic and UI         | `src/features/*/services/` or hooks; functional React UI. Content modules in `src/pages/content/` remain self-contained                                                              |
-| Folders                      | `src/pages/content/folder/`, `src/features/folder/`; persisted types currently exist in both `src/core/types/folder.ts` and `src/pages/content/folder/types.ts`                      |
+| Folders                      | Schema: `src/core/types/folder.ts`; content `types.ts` re-exports it and owns drag data. Logic: `src/features/folder/`; host/UI: `src/pages/content/folder/`                         |
 | Popup settings and shortcuts | `src/pages/popup/Popup.tsx` and its `components/`; shortcuts use `src/core/services/KeyboardShortcutService.ts`, related types and `components/KeyboardShortcutSettings.tsx`         |
 | Cloud sync                   | `src/core/services/GoogleDriveSyncService.ts`                                                                                                                                        |
 | Translations                 | `src/locales/*/messages.json`: all 10 locales for new/removed user-facing keys                                                                                                       |
