+++ src/core/types/folder.ts
@@ -1,13 +1,14 @@
 /**
- * Folder-specific types
- * Extracted from the monolithic FolderManager
+ * Shared folder storage and import/export schema.
+ * Persisted IDs are opaque strings, including legacy and imported IDs.
+ * Normalize conversation identities only at the routing/identity boundary.
  */
 import type { ConversationId, FolderId } from './common';
 
 export interface Folder {
-  readonly id: FolderId;
+  id: string;
   name: string;
-  parentId: FolderId | null;
+  parentId: string | null;
   isExpanded: boolean;
   pinned?: boolean;
   color?: string;
@@ -18,7 +19,7 @@ export interface Folder {
 }
 
 export interface ConversationReference {
-  readonly conversationId: ConversationId;
+  conversationId: string;
   title: string;
   url: string;
   addedAt: number;
