+++ src/features/folder/constants.ts
@@ -1,0 +1,3 @@
+// Keep the serialized keys stable so existing pre-import snapshots remain readable.
+export const SESSION_BACKUP_KEY = 'gvFolderBackup';
+export const SESSION_BACKUP_TIMESTAMP_KEY = 'gvFolderBackupTimestamp';
